Call execHost for communication
