+++ packages/dui3/commands/initLocalAccounts.ts
@@ -1,0 +1,8 @@
+import { Command } from '../types/command'
+
+export type InitLocalAccounts = Command
+
+export const initLocalAccounts: InitLocalAccounts = {
+  name: 'init_local_accounts',
+  data: {}
+}
