+++ packages/dui3/types/callbackFunction.ts
@@ -1,0 +1,3 @@
+import { CallbackMessage } from '~/types/callbackMessage'
+
+export type CallbackFunction = (msg: CallbackMessage) => void
