+++ packages/dui3/types/callbackMessage.ts
@@ -1,0 +1,4 @@
+export type CallbackMessage = {
+  name: string
+  data: object
+}
