+++ packages/dui3/types/command.ts
@@ -1,0 +1,4 @@
+export type Command = {
+  name: string
+  data: object
+}
