+++ packages/dui3/types/global.d.ts
@@ -5,7 +5,7 @@ import { Sketchup } from './sketchup'
 declare global {
   interface Window {
     sketchup: Sketchup
-    loadAccounts: () => void
+    loadAccounts: (accounts: Account[]) => void
     init: () => void
   }
 }
