+++ packages/dui3/types/sketchup.ts
@@ -1,3 +1,5 @@
+import { CallbackMessage } from './callbackMessage'
+
 export type Sketchup = {
-  exec: () => void
+  exec: (msg: CallbackMessage) => void
 }
